+++ src/Nekomata/Particle.hs
@@ -162,7 +162,9 @@ builtinParticles =
         \If the input is an number, apply the function to each integer \
         \from 0 to the input minus 1.\n\
         \If the function takes no argument, return a list of n copies \
-        \of the result of the function, where n is the length of the input."
+        \of the result of the function, where n is the length of the input.\n\
+        \If the function takes more than one argument, the remaining \
+        \arguments are reused unchanged in every recursive call."
         [("[1,2,3] ᵐ{1+}", all_ ["[2,3,4]"])]
     , BuiltinParticle
         "mapWith"
@@ -173,7 +175,9 @@ builtinParticles =
         "Map a binary function over its first argument.\n\
         \If the function is unary, return a list of n copies of the \
         \result of applying the function to the second argument, where \
-        \n is the length of the first argument."
+        \n is the length of the first argument.\n\
+        \If the function takes more than two arguments, the remaining \
+        \arguments are reused unchanged in every recursive call."
         [("[1,2,3] 4 ᵚ{+}", all_ ["[5,6,7]"])]
     , BuiltinParticle
         "zipWith"
@@ -183,7 +187,9 @@ builtinParticles =
         "Zip two lists and apply a function to each pair of elements.\n\
         \Fail if the lists have different lengths.\n\
         \If one of the input is an number, apply the function to each \
-        \integer from 0 to the input minus 1."
+        \integer from 0 to the input minus 1.\n\
+        \If the function takes more than two arguments, the remaining \
+        \arguments are reused unchanged in every recursive call."
         [("[1,2,3] [4,5,6] ᶻ{+}", all_ ["[5,7,9]"])]
     , BuiltinParticle
         "zipWithTrunc"
@@ -194,7 +200,9 @@ builtinParticles =
         \If the lists have different lengths, truncate the longer list \
         \to the length of the shorter list.\n\
         \If one of the input is an number, apply the function to each \
-        \integer from 0 to the input minus 1."
+        \integer from 0 to the input minus 1.\n\
+        \If the function takes more than two arguments, the remaining \
+        \arguments are reused unchanged in every recursive call."
         [("[1,2,3] [4,5,6,7] ᶾ{+}", all_ ["[5,7,9]"])]
     , BuiltinParticle
         "outer"
@@ -204,7 +212,9 @@ builtinParticles =
         "Apply a function to every possible pair of elements in two lists \
         \and return a list of lists.\n\
         \If one of the input is an number, apply the function to each \
-        \integer from 0 to the input minus 1."
+        \integer from 0 to the input minus 1.\n\
+        \If the function takes more than two arguments, the remaining \
+        \arguments are reused unchanged in every recursive call."
         [("[1,2,3] [4,5] ᵒ{+}", all_ ["[[5,6],[6,7],[7,8]]"])]
     , BuiltinParticle
         "concatMap"
@@ -385,6 +395,29 @@ builtinParticles =
         \If the input is an number, it is converted to a list of integers \
         \from 0 to the input minus 1 before applying the function."
         [("[1,2,3] ʰ{1+}", all_ ["[2,2,3]", "[1,3,3]", "[1,2,4]"])]
+    , BuiltinParticle
+        "bottomUp"
+        'ᵇ'
+        bottomUp'
+        "(m -> 1) -> (m -> 1) where m > 0"
+        "Traverse the top value of the stack in a bottom-up manner.\n\
+        \If the value is a list, first apply the function recursively to each \
+        \element, and then apply the function to the resulting list.\n\
+        \If the function takes more than one argument, the remaining \
+        \arguments are reused unchanged in every recursive call."
+        [("[1,[2,3],[[4],5]] ᵇ{1+}", all_ ["[3,[5,6],[[8],8]]"])]
+    , BuiltinParticle
+        "topDown"
+        'ᵀ'
+        topDown'
+        "(m -> 1) -> (m -> 1) where m > 0"
+        "Traverse the top value of the stack in a top-down manner.\n\
+        \First apply the function to the current value.\n\
+        \If the result is a list, recursively apply the function to each \
+        \element of that list.\n\
+        \If the function takes more than one argument, the remaining \
+        \arguments are reused unchanged in every recursive call."
+        [("[1,[2,3],[[4],5]] ᵀ{1+}", all_ ["[2,[4,5],[[7],7]]"])]
     ]
 
 -- | The map of from names to builtin particles
@@ -805,3 +838,25 @@ onAny = Particle onAny'
             (leftId i)
             (Val $ Cons (x >>= f (leftId (rightId i))) xs)
             (Val . Cons x $ xs >>= onAny_ f (rightId (rightId i)))
+
+bottomUp' :: Particle
+bottomUp' = Particle bottomUp''
+  where
+    bottomUp'' (Function (Arity m 1) f) | m > 0 =
+        Just
+            . Function (Arity m 1)
+            $ \i (x :+ s) ->
+                let f' i' x' = top $ f i' (Val x' :+ s)
+                 in (x >>= bottomUp f' i) :+ dropStack (m - 1) s
+    bottomUp'' _ = Nothing
+
+topDown' :: Particle
+topDown' = Particle topDown''
+  where
+    topDown'' (Function (Arity m 1) f) | m > 0 =
+        Just
+            . Function (Arity m 1)
+            $ \i (x :+ s) ->
+                let f' i' x' = top $ f i' (Val x' :+ s)
+                 in (x >>= topDown f' i) :+ dropStack (m - 1) s
+    topDown'' _ = Nothing